# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vending machine management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Clients ───────────────────────────────────────────────────────────────
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateClient
      tags: [clients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /clients/{id}/machines:
    get:
      operationId: listClientMachines
      tags: [clients]
      summary: List machines assigned to a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client machines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientMachine"
    post:
      operationId: assignMachineToClient
      tags: [clients]
      summary: Assign a machine to a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientMachineInput"
      responses:
        "201":
          description: Assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientMachine"
  /clients/{clientId}/machines/{machineId}:
    delete:
      operationId: removeClientMachine
      tags: [clients]
      summary: Remove a machine assignment from a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: machineId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── Machines ──────────────────────────────────────────────────────────────
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Dashboard statistics
      parameters:
        - name: clientId
          in: query
          schema:
            type: ["integer", "null"]
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/top-products:
    get:
      operationId: getDashboardTopProducts
      tags: [dashboard]
      summary: Top and bottom selling products
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Top/bottom products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopProductsResult"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Auth ────────────────────────────────────────────────────────────────
    LoginInput:
      type: object
      properties:
        password:
          type: string
      required:
        - password
    AuthSession:
      type: object
      properties:
        operatorId:
          type: ["integer", "null"]
        operatorName:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, moderator, master]
      required:
        - role

    # ── Client ──────────────────────────────────────────────────────────────
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        hasContract:
          type: boolean
        contractStartDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - hasContract
        - createdAt
    ClientInput:
      type: object
      properties:
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        hasContract:
          type: boolean
        contractStartDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - name
        - hasContract
    ClientUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        hasContract:
          type: ["boolean", "null"]
        contractStartDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ── ClientMachine ────────────────────────────────────────────────────────
    ClientMachine:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        machineId:
          type: integer
        machineName:
          type: string
        machineType:
          type: string
        machineNumber:
          type: string
        installedAt:
          type: ["string", "null"]
        isActive:
          type: boolean
        qrCode:
          type: ["string", "null"]
      required:
        - id
        - clientId
        - clientName
        - machineId
        - machineName
        - machineType
        - machineNumber
        - isActive
    ClientMachineInput:
      type: object
      properties:
        machineId:
          type: integer
        machineNumber:
          type: string
        installedAt:
          type: ["string", "null"]
      required:
        - machineId
        - machineNumber

    # ── Machine ──────────────────────────────────────────────────────────────
...
    DashboardStats:
      type: object
      properties:
        totalRevenue:
          type: number
        totalCost:
          type: number
        totalProfit:
          type: number
        totalExpenses:
          type: number
        netProfit:
          type: number
        totalClients:
          type: integer
        totalMachines:
          type: integer
        totalLoads:
          type: integer
        clientStats:
          type: array
          items:
            $ref: "#/components/schemas/ClientStat"
      required:
        - totalRevenue
        - totalCost
        - totalProfit
        - totalExpenses
        - netProfit
        - totalClients
        - totalMachines
        - totalLoads
        - clientStats
    ClientStat:
      type: object
      properties:
        clientId:
          type: integer
        clientName:
          type: string
        revenue:
          type: number
        cost:
          type: number
        profit:
          type: number
        expenses:
          type: number
        netProfit:
          type: number
        machineCount:
          type: integer
        loadCount:
          type: integer
      required:
        - clientId
        - clientName
        - revenue
        - cost
        - profit
        - expenses
        - netProfit
        - machineCount
        - loadCount
    TopProductsResult:
      type: object
      properties:
        topSelling:
          type: array
          items:
            $ref: "#/components/schemas/ProductStat"
        bottomSelling:
          type: array
          items:
            $ref: "#/components/schemas/ProductStat"
      required:
        - topSelling
        - bottomSelling
    ProductStat:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        totalQuantity:
          type: number
        totalRevenue:
          type: number
        totalProfit:
          type: number
      required:
        - productId
        - productName
        - totalQuantity
        - totalRevenue
        - totalProfit