# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vending machine management API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthSessionRole } from './authSessionRole';

export interface AuthSession {
  /** @nullable */
  operatorId?: number | null;
  /** @nullable */
  operatorName?: string | null;
  role: AuthSessionRole;
}